Pulled constants and dialer logic into separate files.
