+++ daemon/logger/syslog/internal/srslog/srslog.go
@@ -3,65 +3,9 @@ package srslog
 import (
 	"crypto/tls"
 	"crypto/x509"
-	"errors"
 	"io/ioutil"
 	"log"
 	"os"
-)
-
-// The priority is a combination of the syslog facility and
-// severity. For example, LOG_ALERT | LOG_FTP sends an alert severity
-// message from the FTP facility. The default severity is LOG_EMERG;
-// the default facility is LOG_KERN.
-type priority int
-
-const severityMask = 0x07
-const facilityMask = 0xf8
-
-const (
-	// Severity.
-
-	// From /usr/include/sys/syslog.h.
-	// These are the same on Linux, BSD, and OS X.
-	LOG_EMERG priority = iota
-	LOG_ALERT
-	LOG_CRIT
-	LOG_ERR
-	LOG_WARNING
-	LOG_NOTICE
-	LOG_INFO
-	LOG_DEBUG
-)
-
-const (
-	// Facility.
-
-	// From /usr/include/sys/syslog.h.
-	// These are the same up to LOG_FTP on Linux, BSD, and OS X.
-	LOG_KERN priority = iota << 3
-	LOG_USER
-	LOG_MAIL
-	LOG_DAEMON
-	LOG_AUTH
-	LOG_SYSLOG
-	LOG_LPR
-	LOG_NEWS
-	LOG_UUCP
-	LOG_CRON
-	LOG_AUTHPRIV
-	LOG_FTP
-	_ // unused
-	_ // unused
-	_ // unused
-	_ // unused
-	LOG_LOCAL0
-	LOG_LOCAL1
-	LOG_LOCAL2
-	LOG_LOCAL3
-	LOG_LOCAL4
-	LOG_LOCAL5
-	LOG_LOCAL6
-	LOG_LOCAL7
 )
 
 // This interface and the separate syslog_unix.go file exist for
@@ -106,8 +50,8 @@ func DialWithTLSCertPath(network, raddr string, priority priority, tag, certPath
 }
 
 func DialWithTLSConfig(network, raddr string, priority priority, tag string, tlsConfig *tls.Config) (*writer, error) {
-	if priority < 0 || priority > LOG_LOCAL7|LOG_DEBUG {
-		return nil, errors.New("log/syslog: invalid priority")
+	if err := validatePriority(priority); err != nil {
+		return nil, err
 	}
 
 	if tag == "" {
